Commit current state. Unclear what went wrong, but currently, not even the start64.asm is running... (Last commit was better)
To do (except from fixing that nothing works): why does printf not work

+++ scrn.c
@@ -233,7 +233,7 @@ printf (const char *format, ...)
             case 'd':
             case 'u':
             case 'x':
-              itoa (buf, c, *((int *) arg++));
+              itoa (buf, c, *((long *) arg++));
               p = buf;
               goto string;
               break;
@@ -249,9 +249,10 @@ printf (const char *format, ...)
               break;
 
             default:
-              putch (*((int *) arg++));
+              putch (*((long *) arg++));
               break;
             }
         }
     }
 }
+
